cleanup tables fix
